requirements.txt
discord
pixivpy
requests
schedule
more-itertools